adding header panel
