+++ src/App/Layout/AppHeader.jsx
@@ -1,7 +1,7 @@
 import './styles/header.css';
 
 import { ReactComponent as PomeranianLogo } from '../Images/start-it-logo.svg';
-import { HeaderMenu } from '../Components/HeaderMenu';
+import { HeaderMenu } from './components/HeaderMenu';
 
 export function AppHeader() {
   return (
