+++ src/App/Layout/components/LoginPanel/styles.css
@@ -1,0 +1,25 @@
+.header-panel {
+  position: absolute;
+  display: flex;
+  top: 0;
+  right: 0;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: 6px;
+
+  width: 396px;
+  height: 211px;
+  padding: 8px 0px;
+  background-color: yellow;
+  border-radius: 4px;
+  background: #fff;
+
+  box-shadow: 0px 65px 30px 0px #00000003;
+
+  box-shadow: 0px 8px 20px 0px #0000000a;
+
+  box-shadow: 0px 17px 35px 0px #0000000d;
+
+  box-shadow: 0px -4px 26px 0px #00000014;
+}
